pyproject.toml
[tool.poetry]
name = "punctualcoolvirtualmemory"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
request = "^2020.7.1"
beautifulsoup4 = "^4.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
